+++ server/api/index.js
@@ -2,7 +2,8 @@ const router = require('express').Router()
 module.exports = router
 
 router.use('/users', require('./users'));
-// router.use('/flights', require('./flights'));
+router.use('/flights', require('./flights'));
+router.use('/airports', require('./airports'));
 
 router.use((req, res, next) => {
   const error = new Error('Not Found')
